migrate to weakdeps
This adds support for the weakdeps feature on Julia 1.9 and later.
