+++ Project.toml
@@ -26,6 +26,13 @@ Parameters = "0.12"
 Requires = "1"
 julia = "1.4"
 
+[extensions]
+ColorsExt = "Colors"
+DataFramesExt = "DataFrames"
+DistributionsExt = "Distributions"
+IJuliaExt = "IJulia"
+JSON3Ext = "JSON3"
+
 [extras]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
@@ -34,3 +41,10 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
 test = ["Dates", "DelimitedFiles", "Test", "DataFrames"]
+
+[weakdeps]
+Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
