+++ ext/ColorsExt.jl
@@ -1,0 +1,8 @@
+module ColorsExt
+
+isdefined(Base, :get_extension) ? (using Colors) : (using ..Colors)
+using PlotlyBase
+
+_json_lower(a::Colors.Colorant) = string("#", Colors.hex(a))
+
+end
